Fix HalfKAv2 index overflow; add 69M re-encoding support
Critical bug: HalfKAv2 indices reach 45,055 but were stored as int16
(max 32,767). Slots 8-10 (opponent rook/queen/king) were silently
corrupted. The existing 17 shards are recoverable: int16 preserves
the uint16 bit pattern, so a .view(uint16) cast restores original values.

dataset.py (GPUPreloadedDualDataset / ShardedGPUDataset):
- Add feature_dim parameter (default=HALFKP_FEATURE_DIM for backward compat)
- Fix index loading: int16 → .view(uint16).astype(int32) to recover
  HalfKAv2 values; other dtypes cast directly to int32
- SENTINEL now derives from feature_dim instead of hardcoded HALFKP_FEATURE_DIM
- Existing HalfKP shards (uint16 ≤ 24576) are unaffected

train.py:
- Pass feature_dim=cfg["model"]["input_dim"] to ShardedGPUDataset

download_lichess_hf.py:
- Save indices as int32 (was int16); new downloads are overflow-safe

preprocess_dataset.py:
- Add --halfkav2 flag: encodes JSONL to HalfKAv2 binary (45,056-dim)
  using encode_board_halfkav2_dual, idx dtype int32
- Correct count for HalfKAv2: pieces - 1 (own king excluded)

config: document re-encode command for the 69M dataset

To prepare the 69M dataset for HalfKAv2 training:
  PYTHONPATH=. .venv/bin/python3 scripts/preprocess_dataset.py \
    --input data/all_69m/train_all_69m.jsonl \
    --output data/all_69m/train_all_69m \
    --dual --halfkav2
Then uncomment the train_shards list in the config.

+++ nn_training/scripts/download_lichess_hf.py
@@ -70,7 +70,7 @@ def _encode_position(fen: str, cp_stm: float, cp_clamp: float):
     cp_white = float(cp_stm) * sign
     cp_white = float(np.clip(cp_white, -cp_clamp, cp_clamp))
 
-    return w.astype(np.int16), b.astype(np.int16), n_pieces, cp_white, n_pieces
+    return w.astype(np.int32), b.astype(np.int32), n_pieces, cp_white, n_pieces
 
 
 _CP_CLAMP_GLOBAL: float = 3000.0  # set before spawning workers
@@ -145,8 +145,8 @@ def process_shard(
             results = [r for r in pool.map(_encode_row, chunk, chunksize=2048) if r is not None]
             if not results:
                 continue
-            white_parts.append(np.array([r[0] for r in results], dtype=np.int16))
-            black_parts.append(np.array([r[1] for r in results], dtype=np.int16))
+            white_parts.append(np.array([r[0] for r in results], dtype=np.int32))
+            black_parts.append(np.array([r[1] for r in results], dtype=np.int32))
             counts_parts.append(np.array([r[2] for r in results], dtype=np.uint8))
             cp_parts.append(np.array([r[3] for r in results], dtype=np.float32))
             pc_parts.append(np.array([r[4] for r in results], dtype=np.uint8))
